+++ adapters/lumi/sensor_smoke.py
@@ -1,0 +1,8 @@
+from adapters.adapter_with_battery import AdapterWithBattery
+from devices.sensor.smoke import SmokeSensor
+
+
+class SensorSmoke(AdapterWithBattery):
+    def __init__(self, devices):
+        super().__init__(devices)
+        self.devices.append(SmokeSensor(devices, 'smoke', 'smoke'))